requirement.txt
aniso8601==7.0.0
Django==2.2.4
django-environ==0.4.5
django-filter==2.2.0
django-graphql-jwt==0.2.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
graphene==2.1.8
graphene-django==2.5.0
graphene-django-extras==0.4.6
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
py3dns==3.2.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
validate-email==1.3
